requirements.txt
streamlit==1.25.0
pandas==1.5.3
python-dotenv==1.0.0
google-generativeai==0.2.0
